remove redundant story
